projects/capstone/starter/requirements.txt
alembic==1.3.2
apispec==1.3.3
app==0.0.1
argcomplete==1.11.1
astroid==2.2.5
attrs==19.3.0
awscli==1.18.39
Babel==2.8.0
botocore==1.15.39
cached-property==1.5.1
cattrs==0.9.0
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.31
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.15.2
ecdsa==0.13.2
Flask==1.0.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.2.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-OpenID==1.2.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.2
future==0.17.1
graphviz==0.13.2
gunicorn==19.10.0
idna==2.7
importlib-metadata==1.4.0
iso8601==0.1.12
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.5
jose==1.0.0
jsonschema==3.2.0
lazy-object-proxy==1.4.0
lockfile==0.12.2
Mako==1.1.1
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.21.0
mccabe==0.6.1
more-itertools==8.1.0
packaging==20.3
pendulum==1.4.4
pluggy==0.13.1
prison==0.1.2
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycryptodome==3.3.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
pyrant==0.6.5
pyrsistent==0.15.7
pytest==5.4.1
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
python-jose-cryptodome==1.3.2
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3.1
requests==2.20.1
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.3
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.1
tabulate==0.8.6
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
typed-ast==1.3.5
typing-extensions==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.8
virtualenv==16.7.9
wcwidth==0.1.9
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
zipp==2.0.0
zope.deprecation==4.4.0
